+++ miden-crypto/src/dsa/mod.rs
@@ -1,3 +1,4 @@
 //! Digital signature schemes supported by default in the Miden VM.
 
+pub mod ecdsa_k256_keccak;
 pub mod rpo_falcon512;