+++ miden-crypto/src/lib.rs
@@ -7,6 +7,7 @@ extern crate alloc;
 extern crate std;
 
 pub mod dsa;
+pub mod ecdh;
 pub mod hash;
 pub mod merkle;
 pub mod rand;